Prizma

+++ thelegendofbonk/tools.h
@@ -8,8 +8,8 @@
 #define TEXTURE_PATH "..\\assets\\textures\\"
 #define AUDIO_PATH "..\\assets\\audio\\"
 
-#define TICKSPEED 60
-#define TICK_TIME 1 / (float)TICKSPEED
+#define TICKSPEED 60.f
+#define TICK_TIME 1.f / TICKSPEED
 #define PI (float)3.1415926535
 
 #define DEBUG 01
